performance

+++ ft_memcmp.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_memcmp.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: simon <simon@student.42.fr>                +#+  +:+       +#+        */
+/*   By: svan-hoo <svan-hoo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 21:37:21 by svan-hoo          #+#    #+#             */
-/*   Updated: 2023/11/02 16:21:53 by simon            ###   ########.fr       */
+/*   Updated: 2023/11/02 19:21:38 by svan-hoo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,6 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	while (n--)
 		if (*(unsigned char *)s1++ != *(unsigned char *)s2++)
-			return ((int)(*(unsigned char *)(s1 - 1) - *(unsigned char *)(s2 - 1)));
+			return ((*(unsigned char *)(s1 - 1) - *(unsigned char *)(s2 - 1)));
 	return (0);
 }
